Added resource folder with files that have static methods
