+++ .gitignore
@@ -1,2 +1,2 @@
 *.jpg
-*__pycache__/*
+*__pycache__
