+++ Hardcode/run_slopeThresh.sh
@@ -1,0 +1,3 @@
+python ./slopeThreshLED_image.py \
+	-i ../../datasets/RoboMasterLabelledImagesSet1/image-550.jpg \
+	--save ./output/
